repl: remove usage of require('util')
Use  instead of
