+++ lib/internal/modules/esm/module_job.js
@@ -6,7 +6,7 @@ const {
   SafePromise
 } = primordials;
 
-const { decorateErrorStack } = require('internal/util');
+const { decorateErrorStack } = require('internal/util/decorateErrorStack');
 const assert = require('internal/assert');
 const resolvedPromise = SafePromise.resolve();
