+++ lib/internal/process/execution.js
@@ -36,7 +36,7 @@ function tryGetCwd() {
 }
 
 function evalModule(source) {
-  const { decorateErrorStack } = require('internal/util');
+  const { decorateErrorStack } = require('internal/util/decorateErrorStack');
   const asyncESM = require('internal/process/esm_loader');
   asyncESM.loaderPromise.then(async (loader) => {
     const { result } = await loader.eval(source);
