+++ lib/internal/util.js
@@ -14,7 +14,6 @@ const {
   arrow_message_private_symbol: kArrowMessagePrivateSymbolIndex,
   decorated_private_symbol: kDecoratedPrivateSymbolIndex
 } = internalBinding('util');
-const { isNativeError } = internalBinding('types');
 
 const noCrypto = !process.versions.openssl;
 
@@ -24,13 +23,6 @@ const colorRegExp = /\u001b\[\d\d?m/g; // eslint-disable-line no-control-regex
 
 function removeColors(str) {
   return str.replace(colorRegExp, '');
-}
-
-function isError(e) {
-  // An error could be an instance of Error while not being a native error
-  // or could be from a different realm and not be instance of Error but still
-  // be a native error.
-  return isNativeError(e) || e instanceof Error;
 }
 
 function objectToString(o) {
@@ -81,19 +73,6 @@ function deprecate(fn, msg, code) {
   }
 
   return deprecated;
-}
-
-function decorateErrorStack(err) {
-  if (!(isError(err) && err.stack) ||
-      getHiddenValue(err, kDecoratedPrivateSymbolIndex) === true)
-    return;
-
-  const arrow = getHiddenValue(err, kArrowMessagePrivateSymbolIndex);
-
-  if (arrow) {
-    err.stack = arrow + err.stack;
-    setHiddenValue(err, kDecoratedPrivateSymbolIndex, true);
-  }
 }
 
 function assertCrypto() {
@@ -391,14 +370,12 @@ module.exports = {
   assertCrypto,
   cachedResult,
   convertToValidSignal,
-  createClassWrapper,
-  decorateErrorStack,
+  createClassWrapper,  
   deprecate,
   emitExperimentalWarning,
   filterDuplicateStrings,
   getConstructorOf,
-  getSystemErrorName,
-  isError,
+  getSystemErrorName,  
   isInsideNodeModules,
   join,
   normalizeEncoding,
