+++ lib/repl.js
@@ -54,10 +54,10 @@ const {
   isIdentifierChar
 } = require('internal/deps/acorn/acorn/dist/acorn');
 const {
-  decorateErrorStack,
-  isError,
   deprecate
 } = require('internal/util');
+const { isError } = require('internal/util/isError').isError;
+const { decorateErrorStack } = require('internal/util/decorateErrorStack');
 const { inspect } = require('internal/util/inspect');
 const Stream = require('stream');
 const vm = require('vm');
